# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OCMS — Construction ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/milestones:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Milestones]
      operationId: listProjectMilestones
      responses:
        '200':
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Milestone' }
    post:
      tags: [Milestones]
      operationId: createMilestone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MilestoneInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Milestone' }
  /milestones/{milestoneId}:
    parameters:
      - name: milestoneId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Milestones]
      operationId: updateMilestone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MilestoneUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Milestone' }
  /projects/{projectId}/dprs:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [DPR]
      operationId: listProjectDprs
      responses:
        '200':
          description: DPRs
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dpr' }
    post:
      tags: [DPR]
      operationId: createDpr
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DprInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
  /dprs/{dprId}:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [DPR]
      operationId: getDpr
      responses:
        '200':
          description: DPR detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DprDetail' }
  /dprs/{dprId}/submit:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [DPR]
      operationId: submitDpr
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
  /dprs/{dprId}/approve:
    parameters:
      - name: dprId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [DPR]
      operationId: approveDpr
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DprApprovalInput' }
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dpr' }
  /projects/{projectId}/photos:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Photos]
      operationId: listProjectPhotos
      responses:
        '200':
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SitePhoto' }
    post:
      tags: [Photos]
      operationId: createSitePhoto
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SitePhotoInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SitePhoto' }
  /projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Documents]
      operationId: listProjectDocuments
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
    post:
      tags: [Documents]
      operationId: createDocument
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
  /projects/{projectId}/issues:
    parameters:
      - $ref: '#/components/parameters/ProjectIdParam'
    get:
      tags: [Issues]
      operationId: listProjectIssues
      responses:
        '200':
          description: Issues
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Issue' }
    post:
      tags: [Issues]
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
  /issues/{issueId}:
    parameters:
      - name: issueId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Issues]
      operationId: updateIssue
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IssueUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Issue' }
...
  /approvals/{approvalId}/resolve:
    parameters:
      - name: approvalId
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Approvals]
      operationId: resolveApproval
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalResolutionInput' }
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApprovalItem' }
...
components:
  parameters:
...
    ProjectIdParam:
      name: projectId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email: { type: [string, 'null'], format: email }
        firstName: { type: [string, 'null'] }
        lastName: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
...
    UserRole:
      type: string
      enum: [owner, pm, site_engineer, qs, finance, contractor, qc, store, hr, admin]
...
    MilestoneStatus:
      type: string
      enum: [pending, on_track, at_risk, delayed, completed]
    DprStatus:
      type: string
      enum: [draft, submitted, approved, rejected]
    IssueSeverity:
      type: string
      enum: [low, medium, high, critical]
...
    Milestone:
      type: object
      required: [id, projectId, name, description, targetDate, forecastDate, actualDate, status, certificateIssued, sortOrder, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        description: { type: [string, 'null'] }
        targetDate: { type: string, format: date-time }
        forecastDate: { type: [string, 'null'], format: date-time }
        actualDate: { type: [string, 'null'], format: date-time }
        status: { $ref: '#/components/schemas/MilestoneStatus' }
        certificateIssued: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
    MilestoneInput:
      type: object
      required: [name, targetDate]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        targetDate: { type: string, format: date-time }
        forecastDate: { type: string, format: date-time }
        sortOrder: { type: integer }
    MilestoneUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        targetDate: { type: string, format: date-time }
        forecastDate: { type: string, format: date-time }
        actualDate: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/MilestoneStatus' }
        certificateIssued: { type: boolean }
        sortOrder: { type: integer }
    Dpr:
      type: object
      required: [id, projectId, reportDate, weather, temperature, manpowerCount, summary, status, submittedById, approvedById, submittedAt, approvedAt, rejectionReason, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        reportDate: { type: string, format: date-time }
        weather: { type: [string, 'null'] }
        temperature: { type: [number, 'null'] }
        manpowerCount: { type: integer }
        summary: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/DprStatus' }
        submittedById: { type: [string, 'null'] }
        approvedById: { type: [string, 'null'] }
        submittedAt: { type: [string, 'null'], format: date-time }
        approvedAt: { type: [string, 'null'], format: date-time }
        rejectionReason: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    DprInput:
      type: object
      required: [reportDate]
      properties:
        reportDate: { type: string, format: date-time }
        weather: { type: string }
        temperature: { type: number }
        manpowerCount: { type: integer }
        summary: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/DprItemInput' }
    DprApprovalInput:
      type: object
      properties:
        approve: { type: boolean }
        rejectionReason: { type: string }
    DprItemInput:
      type: object
      required: [activityId, quantityToday]
      properties:
        activityId: { type: string }
        quantityToday: { type: number }
        remarks: { type: string }
    DprItem:
      type: object
      required: [id, dprId, activityId, activityName, quantityToday, cumulativeQuantity, remarks]
      properties:
        id: { type: string }
        dprId: { type: string }
        activityId: { type: string }
        activityName: { type: string }
        quantityToday: { type: number }
        cumulativeQuantity: { type: number }
        remarks: { type: [string, 'null'] }
    DprDetail:
      type: object
      required: [dpr, items, photos, submittedBy, approvedBy]
      properties:
        dpr: { $ref: '#/components/schemas/Dpr' }
        items:
          type: array
          items: { $ref: '#/components/schemas/DprItem' }
        photos:
          type: array
          items: { $ref: '#/components/schemas/SitePhoto' }
        submittedBy:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
        approvedBy:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    SitePhoto:
      type: object
      required: [id, projectId, activityId, dprId, url, caption, capturedAt, latitude, longitude, uploadedById, tag, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        activityId: { type: [string, 'null'] }
        dprId: { type: [string, 'null'] }
        url: { type: string }
        caption: { type: [string, 'null'] }
        capturedAt: { type: string, format: date-time }
        latitude: { type: [number, 'null'] }
        longitude: { type: [number, 'null'] }
        uploadedById: { type: [string, 'null'] }
        tag: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    SitePhotoInput:
      type: object
      required: [url]
      properties:
        url: { type: string, minLength: 1 }
        activityId: { type: string }
        dprId: { type: string }
        caption: { type: string }
        capturedAt: { type: string, format: date-time }
        latitude: { type: number }
        longitude: { type: number }
        tag: { type: string }
    Document:
      type: object
      required: [id, projectId, name, category, url, version, uploadedById, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        category: { type: [string, 'null'] }
        url: { type: string }
        version: { type: integer }
        uploadedById: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    DocumentInput:
      type: object
      required: [name, url]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        url: { type: string, minLength: 1 }
        version: { type: integer }
    Issue:
      type: object
      required: [id, projectId, dprId, title, description, severity, status, raisedById, assignedToId, raisedAt, resolvedAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        dprId: { type: [string, 'null'] }
        title: { type: string }
        description: { type: [string, 'null'] }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        status: { type: string }
        raisedById: { type: [string, 'null'] }
        assignedToId: { type: [string, 'null'] }
        raisedAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
    IssueInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        dprId: { type: string }
        assignedToId: { type: string }
    IssueUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        status: { type: string }
        assignedToId: { type: string }
    ApprovalItem:
      type: object
      required: [id, projectId, projectName, entityType, entityId, title, requestedById, requestedByName, assignedToRole, status, createdAt, resolvedAt, ageDays]
      properties:
        id: { type: string }
        projectId: { type: [string, 'null'] }
        projectName: { type: [string, 'null'] }
        entityType: { type: string }
        entityId: { type: string }
        title: { type: string }
        requestedById: { type: [string, 'null'] }
        requestedByName: { type: [string, 'null'] }
        assignedToRole: { $ref: '#/components/schemas/UserRole' }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        ageDays: { type: integer }
    ApprovalResolutionInput:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approved, rejected] }
        comment: { type: string }
...